scheduling included - scheduled job for tracking rejected and cancelled appointments at the end of every day and deleting them
Added scheduled job which will run at the end of every day and check for rejected appointments for that day and delete them. It will also keep track of how many rejections has every employee made and if that number exceeds or is equal to 5 for that day the employee is marked as banned so the admin could later fire him. The method also checks for cancelled appointments and deletes them. A field boolean cancelled is introduced in the appointment entity and is not used the status one so as not to interfere with the count of rejected appointments as a client may decide to cancel his appointment due to it being rejected but this shouldn't change the record of rejected appointments of the employee who did it